+++ instrument-cluster/src/DataManager.hpp
@@ -12,8 +12,8 @@ class DataManager : public QObject {
     // NOLINTNEXTLINE(modernize-use-trailing-return-type)
     Q_OBJECT
     QML_ELEMENT
-    Q_PROPERTY(uint8_t speed READ speed WRITE setSpeed NOTIFY speedChanged)
-    Q_PROPERTY(uint16_t rpm READ rpm WRITE setRpm NOTIFY rpmChanged)
+    Q_PROPERTY(float speed READ speed WRITE setSpeed NOTIFY speedChanged)
+    Q_PROPERTY(float rpm READ rpm WRITE setRpm NOTIFY rpmChanged)
     Q_PROPERTY(uint8_t battery READ battery WRITE setBattery NOTIFY batteryChanged)
     Q_PROPERTY(uint8_t temperature READ temperature WRITE setTemperature NOTIFY temperatureChanged)
 
@@ -21,18 +21,18 @@ class DataManager : public QObject {
     explicit DataManager(QObject* parent = nullptr);
     ~DataManager() override;
 
-    [[nodiscard]] auto speed() const -> uint8_t { return m_speed; }
-    void setSpeed(uint8_t speed);
+    [[nodiscard]] auto speed() const -> float { return m_speed; }
+    void setSpeed(float speed);
 
-    [[nodiscard]] auto rpm() const -> uint16_t { return m_rpm; }
-    void setRpm(uint16_t rpm);
+    [[nodiscard]] auto rpm() const -> float { return m_rpm; }
+    void setRpm(float rpm);
 
     [[nodiscard]] auto battery() const -> uint8_t { return m_battery; }
 
     [[nodiscard]] auto temperature() const -> uint8_t { return m_temperature; }
 
    private slots:
-    void updateData(uint8_t speed, uint16_t rpm);
+    void updateData(float speed, float rpm);
     void setTemperature(uint8_t temperature);
     void setBattery(uint8_t battery);
 
@@ -44,8 +44,8 @@ class DataManager : public QObject {
     void startZeroMQ();
 
    private:
-    uint8_t m_speed = 0;
-    uint16_t m_rpm = 0;
+    float m_speed = 0;
+    float m_rpm = 0;
     uint8_t m_battery = 0;
     uint8_t m_temperature = 0;
     QThread* thread;
